Update circuit.py
With a current source, we need to be careful (at least, I did not know that before I started): the voltage across the current source is independent of the current, therefore it cannot be included in voltage loops, because it assumes the impedance(coefficient) is mulitiplied by the current.  Therefore, we need to remove current sources when computing voltage loops.

Then, we can set a single constraint that a given current (when the source is) is indeed the value expected.